Partial implementation for register + post controller

+++ src/main/java/jcms/repository/RoleRepository.java
@@ -4,7 +4,9 @@ import jcms.model.Role;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.data.rest.core.annotation.RepositoryRestResource;
 
+import java.math.BigInteger;
+
 @RepositoryRestResource(exported = false)
-public interface RoleRepository extends JpaRepository<Role, Long> {
-    Role findById(Long id);
+public interface RoleRepository extends JpaRepository<Role, BigInteger> {
+    Role findById(BigInteger id);
 }
